Safari-609.1.11.1/LayoutTests/http/tests/inspector/network/resource-metrics-expected.txt

Test for Resource metrics (protocol, priority, remoteAddress, connectionIdentifier).


== Running test suite: Resource.Metrics
-- Running test case: Resource.Metrics.Network
PASS: Resource should be created.
PASS: Resource should receive a Response.
PASS: statusCode should be 200.
PASS: protocol should be null or http/1.1.
PASS: priority should be Symbol(unknown) or Symbol(medium).
PASS: remoteAddressType should be null or a string.
PASS: connectionIdentifierType should be null or string.
